requirements.txt
cffi==1.15.1
glfw==2.5.6
multipledispatch==0.6.0
numpy==1.24.2
Pillow==9.4.0
pycparser==2.21
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pyrr==0.10.3
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
